@@ -2,8 +2,9 @@ package com.contoso.demoapp;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import static com.contoso.demoapp.R.id.open_rn_activity;
 
