@@ -6,11 +6,13 @@ import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
 import android.provider.Settings;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.Toast;
 
+import androidx.appcompat.app.AppCompatActivity;
+
 import com.contoso.reactnativedemolibrary.ReactNativeDemoLibrary;
+import com.facebook.soloader.SoLoader;
 
 /**
  * Activity hosting React Native demo.
@@ -28,6 +30,12 @@ public class ReactDemoActivity extends AppCompatActivity {
                     Uri.parse("package:" + getPackageName()));
             startActivityForResult(intent, OVERLAY_PERMISSION_REQ_CODE);
             return;
+        }
+
+        try {
+            SoLoader.init(getApplication(), 0);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
         }
 
         openReactView();
