@@ -1,9 +1,10 @@
 package com.contoso.hostapp1;
 
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 public class MainActivity1 extends AppCompatActivity {
 
