@@ -1,11 +1,13 @@
 package com.contoso.hostapp2;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 import android.widget.FrameLayout;
 
+import androidx.appcompat.app.AppCompatActivity;
+
 import com.contoso.reactnativedemolibrary.ReactNativeDemoLibrary;
+import com.facebook.soloader.SoLoader;
 
 /**
  * Activity showing a normal Android layout that includes the react-native demo view.
@@ -15,9 +17,15 @@ public class MainActivity2 extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main2);
 
-        FrameLayout frame = (FrameLayout) findViewById(R.id.rn_view);
+        try {
+            SoLoader.init(getApplication(), 0);
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+
+        setContentView(R.layout.activity_main2);
+        FrameLayout frame = findViewById(R.id.rn_view);
         View view = ReactNativeDemoLibrary.createHelloWorldView(this, false);
         frame.addView(view);
     }
